implementing into strucutre
